// Compound-Frontend/src/app/services/system-page-actions.enum.ts
export enum SystemPageActions {
  /**
   * إضافة خبر
   */
  NewsAdd = "NewsAdd",
  /**
   * تعديل خبر
   */
  NewsEdit = "NewsEdit",
  /**
   * قائمة الأخبار
   */
  NewsDisplay = "NewsDisplay",
  /**
   * حذف خبر
   */
  NewsDelete = "NewsDelete",
  /**
   * إلغاء النشر الأخبار
   */
  NewsUnpublish = "NewsUnpublish",

  /**
   * إضافة الإشعارات
   */
  NotificationsAdd = "NotificationsAdd",
  /**
   * تعديل الإشعارات
   */
  NotificationsEdit = "NotificationsEdit",
  /**
   * قائمة الإشعارات
   */
  NotificationsDisplay = "NotificationsDisplay",
  /**
   * حذف الإشعارات
   */
  NotificationsDelete = "NotificationsDelete",

  /**
 * إضافة إدارة أدوار المستخدمين
 */
  RolesManagementAdd = "RolesManagementAdd",
  /**
   * تعديل إدارة أدوار المستخدمين
   */
  RolesManagementEdit = "RolesManagementEdit",
  /**
   * قائمة إدارة أدوار المستخدمين
   */
  RolesManagementDisplay = "RolesManagementDisplay",
  /**
   * حذف إدارة أدوار المستخدمين
   */
  RolesManagementDelete = "RolesManagementDelete",

  /**
* إضافة المجمعات
*/
  CompoundsAdd = "CompoundsAdd",
  /**
   * تعديل المجمعات
   */
  CompoundsEdit = "CompoundsEdit",
  /**
   * إدارة المشكلات المجمعات
   */
  CompoundsManageIssue = "CompoundsManageIssue",
  /**
   * حذف المجمعات
   */
  CompoundsDelete = "CompoundsDelete",
  /**
* إدارة الخدمات المجمعات
*/
  CompoundsManageService = "CompoundsManageService",

  /**
* إضافة إعلان
*/
  AdsAdd = "AdsAdd",
  /**
   * تعديل إعلان
   */
  AdsEdit = "AdsEdit",

  /**
   * حذف إعلان
   */
  AdsDelete = "AdsDelete",

  /**
 * قائمة إعلان
 */
  AdsDisplay = "AdsDisplay",


  /**
   *تأكد المستخدمون المسجلون
   */
  RegisteredUsersConfirmed = "RegisteredUsersConfirmed",

  /**
   * إلغاء المستخدمون المسجلون
   */
  RegisteredUsersCancel = "RegisteredUsersCancel",

  /**
 * قائمة المستخدمون المسجلون
 */
  RegisteredUsersDisplay = "RegisteredUsersDisplay",
  /**
    * قائم الوحدات
    */
  UnitsDisplayandFullManagment = "UnitsDisplayandFullManagment",
  /**
  * إضافة البوابات
  */
  GatesAdd = "GatesAdd",
  /**
  * تعديل  البوابات
 */
  GatesEdit = "GatesEdit",

  /**
  * حذف  البوابات
  */
  GatesDelete = "GatesDelete",

  /**
  * قائمة البوابات
  */
  GatesDisplay = "GatesDisplay",
  /**
  *   انجاز الطلب المشكلات
 */
  IssuesMarkasdone = "IssuesMarkasdone",

  /**
  * الغاء الطلب المشكلات
  */
  IssuesCancelRequest = "IssuesCancelRequest",

  /**
  * قائمة المشكلات
  */
  IssuesDisplay = "IssuesDisplay",
  /**
  * تأكد الزيارات
  */
  VisitConfirm = "VisitConfirm",

  /**
  * قائمة الزيارات
  */
  VisitDisplay = "VisitDisplay",
  /**
  *   ادارة الوحدة للملاك
 */
  OwnersManageUnit = "OwnersManageUnit",

  /**
  * حذف الملاك
  */
  OwnersDelete = "OwnersDelete",

  /**
  * إلغاء حساب المستخدم
  */
  OwnersCancelUserAccount = "OwnersCancelUserAccount",
  /**
  * تعديل الملاك
  */
  OwnersEdit = "OwnersEdit",

  /**
  *قائمة الملاك
  */
  OwnersDisplay = "OwnersDisplay",
  /**
 * إضافة المتاجر
 */
  StoresAdd = "StoresAdd",
  /**
  * تعديل المتاجر
  */
  StoresEdit = "StoresEdit",

  /**
  * حذف المتاجر
  */
  StoresDelete = "StoresDelete",
  /**
  * قائمة المتاجر
 */
  StoresDisplay = "StoresDisplay",
  /**
  * غير نشط المتاجر
 */
  StoresDeactive = "StoresDeactive",

  /**
  *  الغاء الطلب الخدمات
  */
  ServicesCancelRequest = "ServicesCancelRequest",
  /**
  *  قائمة الخدمات
 */
  ServicesDisplay = "ServicesDisplay",
  /**
  *  انجاز الطلب الخدمات
 */
  ServicesMarkasdone = "ServicesMarkasdone",
  /**
 * إضافة مستخدمين الإدارة
 */
  ManagementUsersAdd = "ManagementUsersAdd",
  /**
  * تعديل مستخدمين الإدارة
  */
  ManagementUsersEdit = "ManagementUsersEdit",

  /**
  * حذف مستخدمين الإدارة
  */
  ManagementUsersDelete = "ManagementUsersDelete",
  /**
  * قائمة مستخدمين الإدارة
 */
  ManagementUsersDisplay = "ManagementUsersDisplay",
  /**
  * غير نشط مستخدمين الإدارة
 */
  ManagementUsersDeactive = "ManagementUsersDeactive",

}